requirements.txt
fastapi
uvicorn
httpx
pydantic
asyncio
python-dotenv